pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Kerwinson", email = "joshua.hall2020@outlook.com"},
]
dependencies = [
    "twilio>=8.7.0",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "bs4>=0.0.1",
    "psycopg2>=2.9.7",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.97.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts.start]
call = "src.house_alerts.alert:main"